backend/requirements.txt
sanic
sanic_cors
flask_socketio
gevent
